% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBNM.R
\name{BBNM}
\alias{BBNM}
\title{Bayesian Bridge Normal Mean Model}
\usage{
BBNM(
  y,
  step_sizes_tuning_iterations,
  Tb,
  burn,
  nmc,
  thin,
  method.alpha,
  alpha_sig = 0.05
)
}
\arguments{
\item{y}{A numeric vector of the response variable, assumed to be normally distributed, with length p.}

\item{step_sizes_tuning_iterations}{The number of iterations dedicated to tuning the MCMC sampling algorithm for optimal step sizes.}

\item{Tb}{The frequency of tuning adjustments within the initial tuning phase.}

\item{burn}{The number of initial MCMC iterations to discard, allowing the algorithm to reach stationary distribution.}

\item{nmc}{The total number of MCMC iterations to perform post-tuning and burn-in, for final analysis.}

\item{thin}{Thinning interval for the MCMC sampling to reduce autocorrelation in the chain.}

\item{method.alpha}{Specifies the method to update the alpha parameter. It can be "fixed" for a constant alpha throughout the sampling, or "beta" to sample alpha from a Beta distribution.}

\item{alpha_sig}{The significance level for calculating credible intervals for the estimated parameters.}
}
\value{
Returns a list of outputs including the mean estimates of coefficients (BetaHat), credible intervals for these estimates (LeftCI, RightCI), the median of the coefficients (BetaMedian), the estimated error variance (Sigma2Hat), the estimated lambda parameter of the bridge prior (LambdaHat), the estimated alpha parameter of the bridge prior (AlphaHat), and samples from their posterior distributions (BetaSamples, Sigma2Samples, LambdaSamples, AlphaSamples).
}
\description{
This function implements the Bayesian bridge normal mean model, which is an adaptation of the Bayesian bridge linear regression for scenarios where the response variable y follows a normal distribution with mean values influenced by the predictor variables. It is designed to handle high-dimensional datasets and employs a Markov Chain Monte Carlo (MCMC) sampling method for inference.
}
\examples{
\donttest{
# Assuming y is your normally distributed response variable
result <- BBNM(y, tuning=1000, Tb=100, burn=500, nmc=5000, thin=5, method.alpha="beta")
}
}
